doclens/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.3.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
groq==0.9.0
python-dotenv==1.0.1
streamlit==1.35.0
requests==2.32.3
numpy==1.26.4
